Fixes #448: refactor grapheme engine context
